+++ configurations/conemsstate.h
@@ -20,7 +20,7 @@ class ConEMSState
 {
     Q_GADGET
     Q_PROPERTY(QJsonObject currentState READ currentState WRITE setCurrentState USER true)
-    Q_PROPERTY(int timestamp READ timestamp WRITE setTimestamp USER true)
+    Q_PROPERTY(long timestamp READ timestamp WRITE setTimestamp USER true)
 
 public:
     ConEMSState();
@@ -30,15 +30,15 @@ public:
     QJsonObject currentState() const;
     void setCurrentState(const QJsonObject currentState);
 
-    int timestamp() const;
-    void setTimestamp(const int timestamp);
+    long timestamp() const;
+    void setTimestamp(const long timestamp);
 
     bool operator==(const ConEMSState &other) const;
     bool operator!=(const ConEMSState &other) const;
 
 private:
     QJsonObject m_currentState = QJsonObject() ;
-    int m_timestamp = 0;
+    long m_timestamp = 0;
